Creating a quadrilateral class.
Wanted to create a 4-vertex container class. Naturally, I had wanted to create a Rectangle class. I realised that you need to impose restrictions on the Rectangle for it to truly be a rectangle in 3d space. Instead decided to create a "dumber" Quadrilateral class until I decide on something for Rectangles